requirements.txt

streamlit
groq
serpapi
python-dotenv
